Refactor DialogWithCancelButton component
