+++ .moon/tasks.yml
@@ -7,5 +7,5 @@ tasks:
   affected:
     command: noop
     options:
-      cache: false
+      #cache: false
       internal: true
