--- libs/ui-i18n/action/en-US.ts
@@ -1,5 +1,0 @@
-export default {
-  cancel: {
-    label: 'Cancel',
-  },
-};
