--- libs/ui-i18n/action/fr-FR.ts
@@ -1,5 +1,0 @@
-export default {
-  cancel: {
-    label: 'Annuler',
-  },
-};
