--- libs/ui-i18n/action/index.ts
@@ -1,10 +1,0 @@
-import enUs from './en-US';
-import frFr from './fr-FR';
-
-export const messages = {
-  'fr-FR': frFr,
-  'en-US': enUs,
-  'en-GB': enUs,
-};
-
-export const TRANSLATION_KEY = 'lychen_action';
