+++ libs/vue/ui-components-app/dialogs/with-cancel-button/DialogWithCancelButton.vue
@@ -1,0 +1,60 @@
+<template>
+  <Dialog>
+    <DialogTrigger as-child>
+      <slot />
+    </DialogTrigger>
+    <DialogContent class="md:max-w-[30%] w-full max-h-dvh overflow-y-auto gap-8">
+      <DialogHeader>
+        <slot name="header">
+          <DialogTitle>{{ title }}</DialogTitle>
+          <DialogDescription> {{ description }} </DialogDescription>
+        </slot>
+      </DialogHeader>
+      <slot name="content" />
+      <DialogFooter>
+        <slot name="footer">
+          <DialogClose
+            v-if="hasCancelButton"
+            class="flex flex-row justify-end gap-6"
+          >
+            <Button variant="ghost">{{ t('cancel.label') }}</Button>
+
+            <slot name="action" />
+          </DialogClose>
+        </slot>
+      </DialogFooter>
+    </DialogContent>
+  </Dialog>
+</template>
+
+<script lang="ts" setup>
+import Button from '@lychen/vue-ui-components-core/button/Button.vue';
+import {
+  DialogClose,
+  DialogContent,
+  DialogHeader,
+  DialogFooter,
+  DialogTitle,
+  DialogTrigger,
+  Dialog,
+  DialogDescription,
+} from '@lychen/vue-ui-components-core/dialog';
+import { useI18nExtended } from '@lychen/vue-i18n-util-composables/useI18nExtended';
+import { messages, TRANSLATION_KEY } from './i18n';
+
+const {
+  title,
+  description,
+  hasCancelButton = true,
+} = defineProps<{
+  title: string;
+  description: string;
+  hasCancelButton?: boolean;
+}>();
+
+const { t } = useI18nExtended({
+  messages,
+  rootKey: TRANSLATION_KEY,
+  prefixed: true,
+});
+</script>
