+++ libs/vue/ui-components-app/dialogs/with-cancel-button/i18n/en-US.json
@@ -1,0 +1,5 @@
+{
+  "cancel": {
+    "label": "Cancel"
+  }
+}
