+++ libs/vue/ui-components-app/dialogs/with-cancel-button/i18n/fr-FR.json
@@ -1,0 +1,5 @@
+{
+  "cancel": {
+    "label": "Annuler"
+  }
+}
