+++ libs/vue/ui-components-app/dialogs/with-cancel-button/i18n/index.ts
@@ -1,0 +1,10 @@
+import enUs from './en-US.json';
+import frFr from './fr-FR.json';
+
+export const messages = {
+  'fr-FR': frFr,
+  'en-US': enUs,
+  'en-GB': enUs,
+};
+
+export const TRANSLATION_KEY = 'dialog_with_cancel_button';
