+++ libs/vue/ui-components-app/package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@fortawesome/pro-light-svg-icons": "^6.7.2",
     "@lychen/typescript-util-tailwind": "workspace:*",
-    "@lychen/ui-i18n": "workspace:*",
     "@lychen/vue-i18n-util-composables": "workspace:*",
     "@lychen/vue-ui-components-core": "workspace:*",
     "@lychen/vue-ui-components-extra": "workspace:*",
