+++ libs/vue/ui-components-app/tsconfig.json
@@ -2,19 +2,16 @@
   "extends": "../../../tsconfig.options.json",
   "include": [
     "../../typescript/util-tailwind/**/*",
-    "../../ui-i18n/**/*",
     "../i18n/util-composables/**/*",
     "../ui-components-core/**/*",
     "../ui-components-extra/**/*",
     "../util-composables/**/*",
-    "**/*"
+    "**/*",
+    "**/i18n/**/*.json"
   ],
   "references": [
     {
       "path": "../../typescript/util-tailwind"
-    },
-    {
-      "path": "../../ui-i18n"
     },
     {
       "path": "../i18n/util-composables"
@@ -36,8 +33,7 @@
       "@lychen/vue-i18n-util-composables/*": ["../i18n/util-composables/*"],
       "@lychen/vue-ui-components-core/*": ["../ui-components-core/*"],
       "@lychen/vue-ui-components-extra/*": ["../ui-components-extra/*"],
-      "@lychen/vue-util-composables/*": ["../util-composables/*"],
-      "@lychen/ui-i18n/*": ["../../ui-i18n/*"]
+      "@lychen/vue-util-composables/*": ["../util-composables/*"]
     }
   }
 }
